Enum snip1155_reference_impl::msg::HandleMsg

pub enum HandleMsg {

    CurateTokenIds {
        initial_tokens: Vec<CurateTokenId>,
        memo: Option<String>,
        padding: Option<String>,
    },
    MintTokens {
        mint_tokens: Vec<TokenAmount>,
        memo: Option<String>,
        padding: Option<String>,
    },
    BurnTokens {
        burn_tokens: Vec<TokenAmount>,
        memo: Option<String>,
        padding: Option<String>,
    },
    ChangeMetadata {
        token_id: String,
        public_metadata: Box<Option<Metadata>>,
        private_metadata: Box<Option<Metadata>>,
    },
    Transfer {
        token_id: String,
        from: HumanAddr,
        recipient: HumanAddr,
        amount: Uint128,
        memo: Option<String>,
        padding: Option<String>,
    },
    BatchTransfer {
        actions: Vec<TransferAction>,
        padding: Option<String>,
    },
    Send {
        token_id: String,
        from: HumanAddr,
        recipient: HumanAddr,
        recipient_code_hash: Option<String>,
        amount: Uint128,
        msg: Option<Binary>,
        memo: Option<String>,
        padding: Option<String>,
    },
    BatchSend {
        actions: Vec<SendAction>,
        padding: Option<String>,
    },
    GivePermission {
        allowed_address: HumanAddr,
        token_id: String,
        view_balance: Option<bool>,
        view_balance_expiry: Option<Expiration>,
        view_private_metadata: Option<bool>,
        view_private_metadata_expiry: Option<Expiration>,
        transfer: Option<Uint128>,
        transfer_expiry: Option<Expiration>,
        padding: Option<String>,
    },
    RevokePermission {
        token_id: String,
        owner: HumanAddr,
        allowed_address: HumanAddr,
        padding: Option<String>,
    },
    CreateViewingKey {
        entropy: String,
        padding: Option<String>,
    },
    SetViewingKey {
        key: String,
        padding: Option<String>,
    },
    RevokePermit {
        permit_name: String,
        padding: Option<String>,
    },
    AddCurators {
        add_curators: Vec<HumanAddr>,
        padding: Option<String>,
    },
    RemoveCurators {
        remove_curators: Vec<HumanAddr>,
        padding: Option<String>,
    },
    AddMinters {
        token_id: String,
        add_minters: Vec<HumanAddr>,
        padding: Option<String>,
    },
    RemoveMinters {
        token_id: String,
        remove_minters: Vec<HumanAddr>,
        padding: Option<String>,
    },
    ChangeAdmin {
        new_admin: HumanAddr,
        padding: Option<String>,
    },
    RemoveAdmin {
        current_admin: HumanAddr,
        contract_address: HumanAddr,
        padding: Option<String>,
    },
    RegisterReceive {
        code_hash: String,
        padding: Option<String>,
    },

}

Variants

CurateTokenIds

Fields

initial_tokens: Vec<CurateTokenId>

memo: Option<String>

padding: Option<String>

curates new token_ids. Only curators can access this function.

MintTokens

Fields

mint_tokens: Vec<TokenAmount>

memo: Option<String>

padding: Option<String>

BurnTokens

Fields

burn_tokens: Vec<TokenAmount>

memo: Option<String>

padding: Option<String>

ChangeMetadata

Fields

token_id: String

public_metadata: Box<Option<Metadata>>

does not attempt to change if left blank. Can effectively remove metadata by setting metadata to Some(Metadata {token_uri: None, extension: None}) used Box to reduce the total size of the enum variant, to decrease size difference between variants. Not strictly necessary.

private_metadata: Box<Option<Metadata>>

does not attempt to change if left blank. Can effectively remove metadata by setting metadata to Some(Metadata {token_uri: None, extension: None}) used Box to reduce the total size of the enum variant, to decrease size difference between variants. Not strictly necessary.

allows owner or minter to change metadata if allowed by token_id configuration.

Transfer

Fields

token_id: String

from: HumanAddr

recipient: HumanAddr

amount: Uint128

memo: Option<String>

padding: Option<String>

BatchTransfer

Fields

actions: Vec<TransferAction>

padding: Option<String>

Send

Fields

token_id: String

from: HumanAddr

recipient: HumanAddr

recipient_code_hash: Option<String>

amount: Uint128

msg: Option<Binary>

memo: Option<String>

padding: Option<String>

BatchSend

Fields

actions: Vec<SendAction>

padding: Option<String>

GivePermission

Fields

allowed_address: HumanAddr

address being granted/revoked permission

token_id: String

token id to apply approval/revocation to. Additional Spec feature: if == None, perform action for all owner’s token_ids

view_balance: Option<bool>

optional permission level for viewing balance. If ignored, leaves current permission settings

view_balance_expiry: Option<Expiration>

view_private_metadata: Option<bool>

optional permission level for viewing private metadata. If ignored, leaves current permission settings

view_private_metadata_expiry: Option<Expiration>

transfer: Option<Uint128>

set allowance by for transfer approvals. If ignored, leaves current permission settings

transfer_expiry: Option<Expiration>

padding: Option<String>

optional message length padding

RevokePermission

Fields

token_id: String

owner: HumanAddr

token owner

allowed_address: HumanAddr

address which has permission

padding: Option<String>

Removes all permissions that a specific owner has granted to a specific address, for a specific token_id

CreateViewingKey

Fields

entropy: String

padding: Option<String>

SetViewingKey

Fields

key: String

padding: Option<String>

RevokePermit

Fields

permit_name: String

padding: Option<String>

disallow the use of a query permit

AddCurators

Fields

add_curators: Vec<HumanAddr>

padding: Option<String>

RemoveCurators

Fields

remove_curators: Vec<HumanAddr>

padding: Option<String>

AddMinters

Fields

token_id: String

add_minters: Vec<HumanAddr>

padding: Option<String>

RemoveMinters

Fields

token_id: String

remove_minters: Vec<HumanAddr>

padding: Option<String>

ChangeAdmin

Fields

new_admin: HumanAddr

padding: Option<String>

RemoveAdmin

Fields

current_admin: HumanAddr

contract_address: HumanAddr

padding: Option<String>

Permanently breaks admin keys for this contract. No admin function can be called after this action. Any existing curators or minters will remain as curators or minters; no new curators can be added and no current curator can be removed.

Requires caller to input current admin address and contract address. These inputs are not strictly necessary, but as a safety precaution to reduce the chances of accidentally calling this function.

RegisterReceive

Fields

code_hash: String

padding: Option<String>