Function snip1155_reference_impl::state::txhistory::get_txs

pub fn get_txs<S: ReadonlyStorage, A: Api>(
    api: &A, 
    storage: &S, 
    address: &CanonicalAddr, 
    page: u32, 
    page_size: u32
) -> StdResult<(Vec<Tx>, u64)>

Returns StdResult<(Vec, u64)> of the txs to display and the total count of txs

Arguments

api - a reference to the Api used to convert human and canonical addresses

storage - a reference to the contract’s storage

address - a reference to the address whose txs to display

page - page to start displaying

page_size - number of txs per page