Module snip1155_reference_impl::state::permissions

Structs

Permission

struct to store permission for a [token_id, owner, allowed_addr] combination

PermissionKey

to store all keys to access all permissions for a given owner

Functions

list_owner_permission_keys

Return (Vec<PermissionKey { token_id, allowed_addr }>, u64) returns a list and total number of PermissionKeys for a given owner. The PermissionKeys represents (part of) the keys to retrieve all permissions an owner has currently granted

may_load_any_permission

returns StdResult<Option> for a given [owner, token_id, allowed_addr] combination. Returns “dormant” permissions we well, ie: where owner doesn’t currently own tokens. If permission does not exist -> returns StdResult

new_permission

saves new permission entry and adds identifier to the list of permissions the owner address has

update_permission

updates an existing permission entry. Returns error if permission entry does not aleady exist